KWH/requirements.txt
matplotlib
requests
pillow

